requirements.txt
Flask==3.0.0
flasgger==0.9.7.1
tensorflow==2.16.1
numpy
pillow
